+++ internal/table/client.go
@@ -56,7 +56,7 @@ func newClient(
 		index:  make(map[*session]sessionInfo),
 		nodes:  make(map[uint32]map[*session]struct{}),
 		idle:   list.New(),
-		waitq:  list.New(),
+		waitQ:  list.New(),
 		limit:  config.SizeLimit(),
 		waitChPool: sync.Pool{
 			New: func() interface{} {
@@ -92,7 +92,7 @@ type Client struct {
 	createInProgress  int        // KIKIMR-9163: in-create-process counter
 	limit             int        // Upper bound for Client size.
 	idle              *list.List // list<*session>
-	waitq             *list.List // list<*chan *session>
+	waitQ             *list.List // list<*chan *session>
 	waitChPool        sync.Pool
 	testHookGetWaitCh func() // nil except some tests.
 	wg                sync.WaitGroup
@@ -148,6 +148,11 @@ func (c *Client) updateNodes(ctx context.Context, endpoints []endpoint.Info) {
 						}
 					}(s)
 				}
+			}
+		}
+		for _, nodeID := range nodeIDs {
+			if _, ok := c.nodes[nodeID]; !ok {
+				c.nodes[nodeID] = make(map[*session]struct{})
 			}
 		}
 	})
@@ -259,27 +264,29 @@ func (c *Client) createSession(ctx context.Context, opts ...createSessionOption)
 func (c *Client) appendSessionToNodes(s *session) {
 	c.mu.WithLock(func() {
 		nodeID := s.NodeID()
-		sessions, has := c.nodes[nodeID]
-		if !has {
-			sessions = make(map[*session]struct{})
+		if _, has := c.nodes[nodeID]; has {
+			c.nodes[nodeID][s] = struct{}{}
 		}
-		sessions[s] = struct{}{}
-		c.nodes[nodeID] = sessions
 	})
+}
+
+func (c *Client) hasNodeID(nodeID uint32) bool {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	_, ok := c.nodes[nodeID]
+	return ok
 }
 
 func (c *Client) removeSessionFromNodes(s *session) {
 	c.mu.WithLock(func() {
 		nodeID := s.NodeID()
-		sessions, has := c.nodes[nodeID]
-		if !has {
-			sessions = make(map[*session]struct{})
-		}
-		delete(sessions, s)
-		if len(sessions) == 0 {
-			delete(c.nodes, nodeID)
-		} else {
-			c.nodes[nodeID] = sessions
+		if sessions, has := c.nodes[nodeID]; has {
+			delete(sessions, s)
+			if len(sessions) == 0 {
+				delete(c.nodes, nodeID)
+			} else {
+				c.nodes[nodeID] = sessions
+			}
 		}
 	})
 }
@@ -528,7 +535,7 @@ func (c *Client) internalPoolWaitFromCh(ctx context.Context, t trace.Table) (s *
 
 	c.mu.WithLock(func() {
 		ch = c.internalPoolGetWaitCh()
-		el = c.waitq.PushBack(ch)
+		el = c.waitQ.PushBack(ch)
 	})
 
 	waitDone := trace.TableOnPoolWait(t, &ctx)
@@ -545,7 +552,7 @@ func (c *Client) internalPoolWaitFromCh(ctx context.Context, t trace.Table) (s *
 	select {
 	case <-c.done:
 		c.mu.WithLock(func() {
-			c.waitq.Remove(el)
+			c.waitQ.Remove(el)
 		})
 		return nil, xerrors.WithStackTrace(errClosedClient)
 
@@ -566,13 +573,13 @@ func (c *Client) internalPoolWaitFromCh(ctx context.Context, t trace.Table) (s *
 
 	case <-createSessionTimeoutCh:
 		c.mu.WithLock(func() {
-			c.waitq.Remove(el)
+			c.waitQ.Remove(el)
 		})
 		return nil, nil
 
 	case <-ctx.Done():
 		c.mu.WithLock(func() {
-			c.waitq.Remove(el)
+			c.waitQ.Remove(el)
 		})
 		return nil, xerrors.WithStackTrace(ctx.Err())
 	}
@@ -608,6 +615,9 @@ func (c *Client) Put(ctx context.Context, s *session) (err error) {
 
 	case s.isClosed():
 		return xerrors.WithStackTrace(errSessionClosed)
+
+	case !c.hasNodeID(s.NodeID()):
+		return xerrors.WithStackTrace(errNodeIsNotObservable)
 
 	default:
 		c.mu.Lock()
@@ -649,7 +659,7 @@ func (c *Client) Close(ctx context.Context) (err error) {
 
 			c.limit = 0
 
-			for el := c.waitq.Front(); el != nil; el = el.Next() {
+			for el := c.waitQ.Front(); el != nil; el = el.Next() {
 				ch := el.Value.(*chan *session)
 				close(*ch)
 			}
@@ -813,7 +823,7 @@ func (c *Client) internalPoolRemoveFirstIdle() *session {
 
 // c.mu must be held.
 func (c *Client) internalPoolNotify(s *session) (notified bool) {
-	for el := c.waitq.Front(); el != nil; el = c.waitq.Front() {
+	for el := c.waitQ.Front(); el != nil; el = c.waitQ.Front() {
 		// Some goroutine is waiting for a session.
 		//
 		// It could be in this states:
@@ -827,7 +837,7 @@ func (c *Client) internalPoolNotify(s *session) (notified bool) {
 		// missed something and may want to retry (especially for case (3)).
 		//
 		// After that we taking a next waiter and repeat the same.
-		ch := c.waitq.Remove(el).(*chan *session)
+		ch := c.waitQ.Remove(el).(*chan *session)
 		select {
 		case *ch <- s:
 			// Case (1).