+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed bug with put session from unknown node
 * Fixed bug with parsing of `TzTimestamp` without microseconds
 * Fixed code -1 of retryable error if wrapped error with code
 * Added `ydb.WithCompression()` option